Fix sanity check failure because containerd dir does not exist
The update is meant to fix an error on this check when containerd install does not create the /etc/containerd folder. Had this problem while running the "all_install.yml" playbook. Creating the directory and re-running the command makes the trick. 

More info:
VM on Hetzner Cloud - 22.04 5.15.0-56-generic #62-Ubuntu
Containerd - 1.5.9-0ubuntu3.1